SinkBuffer maintain send window
