+++ be/src/exec/pipeline/exchange/sink_buffer.h
@@ -7,19 +7,16 @@
 #include <list>
 #include <mutex>
 #include <queue>
-#include <set>
-#include <unordered_map>
+#include <unordered_set>
 
 #include "column/chunk.h"
 #include "gen_cpp/BackendService.h"
 #include "runtime/current_thread.h"
 #include "runtime/runtime_state.h"
-#include "util/blocking_queue.hpp"
 #include "util/brpc_stub_cache.h"
 #include "util/defer_op.h"
 #include "util/disposable_closure.h"
 #include "util/phmap/phmap.h"
-#include "util/priority_thread_pool.hpp"
 
 namespace starrocks::pipeline {
 
@@ -37,206 +34,26 @@ struct TransmitChunkInfo {
 
 struct ClosureContext {
     TUniqueId instance_id;
-    int64_t request_seq;
+    int64_t sequence;
 };
 
 // TODO(hcf) how to export brpc error
 class SinkBuffer {
 public:
-    SinkBuffer(RuntimeState* state, const std::vector<TPlanFragmentDestination>& destinations, size_t num_sinkers)
-            : _mem_tracker(state->instance_mem_tracker()),
-              _brpc_timeout_ms(std::min(3600, state->query_options().query_timeout) * 1000),
-              _num_uncancelled_sinkers(num_sinkers) {
-        for (const auto& dest : destinations) {
-            const auto& instance_id = dest.fragment_instance_id;
+    SinkBuffer(RuntimeState* state, const std::vector<TPlanFragmentDestination>& destinations, size_t num_sinkers);
+    ~SinkBuffer();
 
-            auto it = _num_sinkers.find(instance_id.lo);
-            if (it != _num_sinkers.end()) {
-                it->second += num_sinkers;
-            } else {
-                _num_sinkers[instance_id.lo] = num_sinkers;
-
-                _request_seqs[instance_id.lo] = 0;
-                _finished_request_seqs[instance_id.lo] = std::set<int64_t>();
-                _buffers[instance_id.lo] = std::queue<TransmitChunkInfo, std::list<TransmitChunkInfo>>();
-                _num_finished_rpcs[instance_id.lo] = 0;
-                _num_in_flight_rpcs[instance_id.lo] = 0;
-                _mutexes[instance_id.lo] = std::make_unique<std::mutex>();
-
-                PUniqueId finst_id;
-                finst_id.set_hi(instance_id.hi);
-                finst_id.set_lo(instance_id.lo);
-                _instance_id2finst_id[instance_id.lo] = std::move(finst_id);
-            }
-
-            _num_remaining_eos = 0;
-            for (auto& [_, num] : _num_sinkers) {
-                _num_remaining_eos += num;
-            }
-        }
-    }
-
-    ~SinkBuffer() {
-        DCHECK(is_finished());
-
-        for (auto& [_, buffer] : _buffers) {
-            while (!buffer.empty()) {
-                auto& request = buffer.front();
-                request.params->release_finst_id();
-                buffer.pop();
-            }
-        }
-    }
-
-    void add_request(const TransmitChunkInfo& request) {
-        DCHECK(_num_remaining_eos > 0);
-        if (_is_finishing) {
-            request.params->release_finst_id();
-            return;
-        }
-        {
-            auto& instance_id = request.fragment_instance_id;
-            std::lock_guard<std::mutex> l(*_mutexes[instance_id.lo]);
-            _buffers[instance_id.lo].push(request);
-            _try_to_send_rpc(instance_id);
-        }
-    }
-
-    bool is_full() const {
-        // std::queue' read is concurrent safe without mutex
-        // Judgement may not that accurate because we do not known in advance which
-        // instance the data to be sent corresponds to
-        size_t max_buffer_size = config::pipeline_sink_buffer_size * _buffers.size();
-        size_t buffer_size = 0;
-        for (auto& [_, buffer] : _buffers) {
-            buffer_size += buffer.size();
-        }
-        return buffer_size > max_buffer_size;
-    }
-
-    bool is_finished() const {
-        if (!_is_finishing) {
-            return false;
-        }
-
-        return _num_sending_rpc == 0 && _total_in_flight_rpc == 0;
-    }
+    void add_request(const TransmitChunkInfo& request);
+    bool is_full() const;
+    bool is_finished() const;
 
     // When all the ExchangeSinkOperator shared this SinkBuffer are cancelled,
     // the rest chunk request and EOS request needn't be sent anymore.
-    void cancel_one_sinker() {
-        if (--_num_uncancelled_sinkers == 0) {
-            _is_finishing = true;
-        }
-    }
+    void cancel_one_sinker();
 
 private:
-    void _try_to_send_rpc(const TUniqueId& instance_id) {
-        DeferOp decrease_defer([this]() { --_num_sending_rpc; });
-        ++_num_sending_rpc;
-
-        for (;;) {
-            if (_is_finishing) {
-                return;
-            }
-
-            auto& buffer = _buffers[instance_id.lo];
-            if (buffer.empty() || _num_in_flight_rpcs[instance_id.lo] >= config::pipeline_sink_brpc_dop) {
-                return;
-            }
-
-            TransmitChunkInfo request = buffer.front();
-            bool need_wait = false;
-            DeferOp pop_defer([&need_wait, &buffer]() {
-                if (need_wait) {
-                    return;
-                }
-                buffer.pop();
-            });
-
-            // The order of data transmiting in IO level may not be strictly the same as
-            // the order of submitting data packets
-            // But we must guarantee that first packet must be received first
-            if (_num_finished_rpcs[instance_id.lo] == 0 && _num_in_flight_rpcs[instance_id.lo] > 0) {
-                need_wait = true;
-                return;
-            }
-            if (request.params->eos()) {
-                DeferOp eos_defer([this, &instance_id, &need_wait]() {
-                    if (need_wait) {
-                        return;
-                    }
-                    if (--_num_remaining_eos == 0) {
-                        _is_finishing = true;
-                    }
-                    --_num_sinkers[instance_id.lo];
-                });
-                // Only the last eos is sent to ExchangeSourceOperator. it must be guaranteed that
-                // eos is the last packet to send to finish the input stream of the corresponding of
-                // ExchangeSourceOperator and eos is sent exactly-once.
-                if (_num_sinkers[instance_id.lo] > 1) {
-                    if (request.params->chunks_size() == 0) {
-                        request.params->release_finst_id();
-                        continue;
-                    } else {
-                        request.params->set_eos(false);
-                    }
-                } else {
-                    // The order of data transmiting in IO level may not be strictly the same as
-                    // the order of submitting data packets
-                    // But we must guarantee that eos packent must be the last packet
-                    if (_num_in_flight_rpcs[instance_id.lo] > 0) {
-                        need_wait = true;
-                        return;
-                    }
-                }
-            }
-
-            request.params->set_allocated_finst_id(&_instance_id2finst_id[instance_id.lo]);
-            request.params->set_sequence(_request_seqs[instance_id.lo]++);
-
-            auto* closure = new DisposableClosure<PTransmitChunkResult, ClosureContext>(
-                    {instance_id, request.params->sequence()});
-            closure->addFailedHandler([this, closure](const ClosureContext& ctx) noexcept {
-                _is_finishing = true;
-                {
-                    std::lock_guard<std::mutex> l(*_mutexes[ctx.instance_id.lo]);
-                    ++_num_finished_rpcs[ctx.instance_id.lo];
-                    --_num_in_flight_rpcs[ctx.instance_id.lo];
-                }
-                --_total_in_flight_rpc;
-                LOG(WARNING) << " transmit chunk rpc failed";
-            });
-            closure->addSuccessHandler(
-                    [this, closure](const ClosureContext& ctx, const PTransmitChunkResult& result) noexcept {
-                        Status status(result.status());
-                        {
-                            std::lock_guard<std::mutex> l(*_mutexes[ctx.instance_id.lo]);
-                            ++_num_finished_rpcs[ctx.instance_id.lo];
-                            --_num_in_flight_rpcs[ctx.instance_id.lo];
-                        }
-                        if (!status.ok()) {
-                            _is_finishing = true;
-                            LOG(WARNING) << " transmit chunk rpc failed, " << status.message();
-                        } else {
-                            std::lock_guard<std::mutex> l(*_mutexes[ctx.instance_id.lo]);
-                            _try_to_send_rpc(ctx.instance_id);
-                        }
-                        --_total_in_flight_rpc;
-                    });
-
-            ++_total_in_flight_rpc;
-            ++_num_in_flight_rpcs[instance_id.lo];
-
-            closure->cntl.Reset();
-            closure->cntl.set_timeout_ms(_brpc_timeout_ms);
-            closure->cntl.request_attachment().append(request.attachment);
-            request.brpc_stub->transmit_chunk(&closure->cntl, request.params.get(), &closure->result, closure);
-
-            request.params->release_finst_id();
-            return;
-        }
-    }
+    void _process_send_window(const TUniqueId& instance_id, const int64_t sequence);
+    void _try_to_send_rpc(const TUniqueId& instance_id);
 
     MemTracker* _mem_tracker = nullptr;
     const int32_t _brpc_timeout_ms;
@@ -246,8 +63,12 @@ private:
     // because TUniqueId::hi is exactly the same in one query
 
     phmap::flat_hash_map<int64_t, int64_t> _num_sinkers;
-    phmap::flat_hash_map<int64_t, int64_t> _request_seqs;
-    phmap::flat_hash_map<int64_t, std::set<int64_t>> _finished_request_seqs;
+    phmap::flat_hash_map<int64_t, int64_t> _request_sequences;
+    // distribution of received response sequence numbers:
+    // part1: { seq | 1 <= seq <= _max_processed_sequence }
+    // part2: { seq | seq = _max_processed_sequence + i, i > 1 }
+    phmap::flat_hash_map<int64_t, int64_t> _max_processed_sequences;
+    phmap::flat_hash_map<int64_t, std::unordered_set<int64_t>> _unprocessed_sequences;
     std::atomic<int32_t> _total_in_flight_rpc = 0;
     std::atomic<int32_t> _num_uncancelled_sinkers;
     std::atomic<int32_t> _num_remaining_eos = 0;
